requirements.txt
pytest==7.4.0
pytest-asyncio==0.21.1
python-telegram-bot==20.4
gspread==5.10.0
oauth2client==4.1.3
python-dotenv==1.0.0
aiohttp>=3.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0